Extract error enums to a separate file
Starting to better organize things.
